Move velocity from the CombatAttributes to IMobileEntity
Issue #89

+++ src/de/gurkenlabs/litiengine/annotation/MovementInfo.java
@@ -22,5 +22,5 @@ public @interface MovementInfo {
    *
    * @return the velocity
    */
-  short velocity() default 100;
+  float velocity() default 100;
 }
